# openapi.yaml
...
openapi: 3.1.0
info:
  title: Traccar
  description: Traccar GPS tracking server API documentation. To use the API you need to have a server instance. For testing purposes you can use one of free [demo servers](https://www.traccar.org/demo-server/). For production use you can install your own server or get a [subscription service](https://www.traccar.org/product/tracking-server/).
  contact:
    name: Traccar Support
    url: https://www.traccar.org/
    email: support@traccar.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '6.14.4'
servers:
  - url: https://demo.traccar.org/api
    description: Demo Server 1
  - url: https://demo2.traccar.org/api
    description: Demo Server 2
  - url: https://demo3.traccar.org/api
    description: Demo Server 3
  - url: https://demo4.traccar.org/api
    description: Demo Server 4
  - url: https://server.traccar.org/api
    description: Subscription Server
  - url: http://{host}:{port}/api
    description: Other Server
    variables:
      host:
        default: localhost
      port:
        enum:
          - '8082'
          - '80'
        default: '8082'
security:
  - BasicAuth: []
  - ApiKey: []
...
paths:
...
  /driver-scripts/{id}/enable:
    post:
      summary: Enable a driver script hash
      description: Admin-only approval for the exact registered filename and SHA-256 hash.
      tags:
        - Driver Scripts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverScript'
        '404':
          description: Not Found
          content: {}
  /driver-scripts/{id}/disable:
    post:
      summary: Disable a driver script hash
      description: Admin-only removal of approval for a registered driver script hash.
      tags:
        - Driver Scripts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverScript'
        '404':
          description: Not Found
          content: {}
...
components:
  schemas:
...
    DriverScript:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique driver script registration identifier
        fileName:
          type: string
          description: Driver script file name
        hash:
          type: string
          description: SHA-256 hash of the registered script contents
        enabled:
          type: boolean
          description: Whether this exact filename and hash is approved for execution
        discoveredTime:
          type: string
          format: date-time
          description: Time when this script hash was first discovered
        lastSeenTime:
          type: string
          format: date-time
          description: Time when this script hash was most recently seen on disk
        loadedTime:
          type: string
          format: date-time
          description: Time when this script hash was last loaded successfully
        error:
          type: string
          description: Last load error for this script hash
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKey:
      type: http
      scheme: bearer